Allow line number of reloading loops to change between reloads
Before, the loop body was always located by line number. That leads to
errors if the source file is edited in such a way that the line number
of the loop changes between reloads.

Now, the loop is located by line number on the first reload. We then
compute a loop identifier from the loop source (similar to simply using the
source code string of the loop line) and find the loop body by that
identifier in subsequent reloads. We're using a `loop_id` instead of the
exact source string as older versions of Python have no way to go from
ast to source string. Instead, we use `ast.dump` which is equivalent to
the source string, but without being valid Python code.