Add extra checks for get_decorator_name()
For some of our decorators, `dec_node.func.id` was raising an exception
since `id` was not an attribute on `func`. This adds a check for it,
as well as a check for `dec_node.func.value.id`, which is where the name
of our decorator was.

This also changes the name to get_decorator_name_or_none() to emphasize
that it may return `None` if it doesn't find an id anywhere.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>